[BAD-1281]-Develop and DEV Validate the MapR 6.0 Shim against Pentaho 8.1 (Master) EE/Secured cluster
