+++ api/src/main/java/org/pentaho/hbase/shim/spi/HBaseConnection.java
@@ -43,6 +43,8 @@ public abstract class HBaseConnection {
   public static final String SITE_KEY = "hbase.site";
   public static final String ZOOKEEPER_QUORUM_KEY = "hbase.zookeeper.quorum";
   public static final String ZOOKEEPER_PORT_KEY = "hbase.zookeeper.property.clientPort";
+  //constant for active shim configuration id
+  public static final String ACTIVE_SHIM_VERSION = "ACTIVE_SHIM_VERSION";
 
   // constant table creation option keys (commented out keys don't exist as
   // options for HColumnDescriptor in 0.90.3)
