+++ common/common-shim/src/main/resources/org/pentaho/hbase/shim/common/messages/messages_en_US.properties
@@ -10,4 +10,5 @@ CommonHBaseConnection.Error.AtEndOfResultSet=At end of result set
 CommonHBaseConnection.Error.ObjectIsNotAnHBaseRow=The supplied object is not an HBase row object
 CommonHBaseConnection.Error.NoTargetTable=No target table has been specified
 CommonHBaseConnection.Error.NoTargetPut=No target table put has been specified
+CommonHBaseConnection.Error.HBaseMapR60SpecificPropertyNotSet=Property hbase.table.namespace.mappings not set. For MapR 6.0 or above you should set that property. More information - https://maprdocs.mapr.com/home/UpgradeGuide/MappingTableNamespace-HBase-DBbinary.html?hl=hbase
