--- shims/mapr510/assemblies/mapr510-shim/src/assembly/assembly.xml
@@ -1,78 +1,0 @@
-<assembly>
-  <id>plugin</id>
-  <baseDirectory>mapr510</baseDirectory>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <fileSets>
-    <fileSet>
-      <directory>${basedir}/src/main/resources</directory>
-      <outputDirectory>/</outputDirectory>
-    </fileSet>
-    <fileSet>
-      <directory>${oss.license.directory}</directory>
-      <outputDirectory>/</outputDirectory>
-    </fileSet>
-  </fileSets>
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>/</outputDirectory>
-      <useTransitiveDependencies>false</useTransitiveDependencies>
-      <useTransitiveFiltering>false</useTransitiveFiltering>
-      <includes>
-        <include>org.pentaho:pentaho-hadoop-shims-mapr510</include>
-        <include>org.pentaho:pentaho-hadoop-shims-mapr510-hbase-comparators</include>
-      </includes>
-    </dependencySet>
-    <dependencySet>
-      <outputDirectory>lib</outputDirectory>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <useTransitiveFiltering>false</useTransitiveFiltering>
-      <includes>
-	    <include>org.xerial.snappy:snappy-java</include>
-        <include>dk.brics.automaton:automaton</include>
-        <include>org.apache.hive:hive-common</include>
-        <include>org.apache.hive:hive-exec</include>
-        <include>org.apache.hive:hive-jdbc</include>
-        <include>org.apache.hive:hive-metastore</include>
-        <include>org.apache.hive:hive-service</include>
-        <include>org.apache.hive.shims:hive-shims-0.23</include>
-        <include>org.apache.hive.shims:hive-shims-common</include>
-        <include>org.apache.thrift:libfb303</include>
-        <include>org.apache.thrift:libthrift</include>
-        <include>com.yammer.metrics:metrics-core</include>
-        <include>org.apache.oozie:oozie-client</include>
-        <include>org.apache.oozie:oozie-core</include>
-        <include>pentaho:pentaho-hdfs-vfs</include>
-        <include>org.apache.pig:pig</include>
-        <include>org.apache.sqoop:sqoop</include>
-        <include>com.twitter:parquet-hadoop-bundle</include>
-        <include>org.apache.avro:avro</include>
-        <include>com.fasterxml.jackson.core:jackson-annotations</include>
-        <include>com.fasterxml.jackson.core:jackson-core</include>
-        <include>com.fasterxml.jackson.core:jackson-databind</include>
-      </includes>
-      <excludes>
-        <exclude>*:tests:*</exclude>
-      </excludes>
-    </dependencySet>
-    <dependencySet>
-      <outputDirectory>lib/pmr</outputDirectory>
-      <useTransitiveDependencies>true</useTransitiveDependencies>
-      <useTransitiveFiltering>false</useTransitiveFiltering>
-      <includes>
-        <include>org.apache.htrace:htrace-core</include>
-        <include>org.apache.zookeeper:zookeeper</include>
-        <include>org.apache.hbase:hbase-client</include>
-        <include>org.apache.hbase:hbase-common</include>
-        <include>org.apache.hbase:hbase-hadoop-compat</include>
-        <include>org.apache.hbase:hbase-protocol</include>
-        <include>org.apache.hbase:hbase-server</include>
-        <include>net.sf.flexjson:flexjson</include>
-      </includes>
-      <excludes>
-        <exclude>*:tests:*</exclude>
-      </excludes>
-    </dependencySet>
-  </dependencySets>
-</assembly>
