--- shims/mapr510/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.pentaho</groupId>
-    <artifactId>pentaho-hadoop-shims-list</artifactId>
-    <version>8.1.0.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>pentaho-hadoop-shims-mapr510-reactor</artifactId>
-  <version>8.1.0.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-  <modules>
-    <module>impl</module>
-    <module>hbase-comparators</module>
-    <module>assemblies</module>
-  </modules>
-  <properties>
-    <org.apache.hadoop.version>2.7.0-mapr-1602</org.apache.hadoop.version>
-    <shim.name>mapr510</shim.name>
-  </properties>
-</project>
