+++ shims/mapr60/assemblies/mapr60-shim/src/main/resources/config.properties
@@ -1,0 +1,44 @@
+# Friendly name for this configuration
+name=MapR 6.0
+
+# Comma-separated list of directories and files to make available to this
+# configuration on linux. Any resources found here will overwrite ones in lib/.
+# Current classpath was received from running "hadoop classpath" and concating ${PENTAHO_INSTALLED_DIR}/design-tools/data-integration/plugins/pentaho-big-data-plugin/hadoop-configurations/mapr60,/opt/mapr/lib
+# For example:
+linux.classpath=/opt/mapr/hadoop/hadoop-2.7.0/etc/hadoop,/opt/mapr/hadoop/hadoop-2.7.0/etc/hadoop,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/common/lib,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/common,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/hdfs,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/hdfs/lib,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/yarn/lib,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/yarn,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/mapreduce/lib,/opt/mapr/hadoop/hadoop-2.7.0/share/hadoop/mapreduce,/opt/mapr/sqoop/sqoop-1.4.6,/opt/mapr/sqoop/sqoop-1.4.6/lib,/contrib/capacity-scheduler,/opt/Pentaho/design-tools/data-integration/plugins/pentaho-big-data-plugin/hadoop-configurations/mapr60,/opt/mapr/lib
+
+
+# Comma-separated list of paths that contain native libraries to load. These
+# could be added to LD_LIBRARY_PATH or set with -Djava.library.path instead.
+linux.library.path=/opt/mapr/lib
+
+# Comma-separated list of classes or package names to explicitly ignore when
+# loading classes from the resources within this Hadoop configuration directory
+# or the classpath property
+# e.g.: org.apache.commons.log,org.apache.log4j
+# Note, the two packages above are automatically included for all configurations
+ignore.classes=
+
+# Comma-separated list of jars to explicitly ignore when
+# loading classes from the resources within this Hadoop configuration directory
+# or the classpath property
+# e.g.: without versions - xercesImpl,xml-apis or with versions - xercesImpl-2.9.1,xml-apis-1.3.04 or
+# with jar extension - xercesImpl-2.9.1.jar,xml-apis-1.3.04.jar
+# Note, the two jars above lead to libraries conflicts on Mapr 6.0 cluster so they are added to exclude.jars property below
+exclude.jars=xercesImpl,xml-apis
+
+# These are Windows-specific classpath and library paths. 
+# Please make sure to update the MapR versions in the paths to match your
+# locally installed MapR client.
+# Make sure you have properly configured windows.classpath property (if you are using PDI on windows): The first entrance should be relative path to lib/hadoop2-windows-patch-08072014.jar
+windows.classpath=file:///C:/opt/mapr/conf,file:///C:/opt/mapr/hadoop/hadoop-0.20.2/lib
+windows.library.path=C:\\opt\\mapr\\lib
+
+java.system.hadoop.login=hybrid
+java.system.mapr.library.flatclass=true
+
+# These classes will be instantiated using Class.forName().newInstance().
+activator.classes=
+
+# Java will shutdown if these classes are not available.
+#required.classes=org.apache.hadoop.HadoopVersionAnnotation,org.apache.hadoop.security.authentication.client.AuthenticatedURL,com.mapr.log4j.CentralTaskLogAppender,mapr.fs.Replicationserver$InodeAllocationBmapMsgOrBuilder,com.mapr.fs.ShimLoader
