+++ shims/mapr60/assemblies/mapr60-shim/src/main/resources/core-site.xml
