+++ shims/mapr60/default/src/assembly/assembly.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly>
+  <id>plugin</id>
+  <baseDirectory>mapr60</baseDirectory>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>lib</outputDirectory>
+      <useTransitiveDependencies>true</useTransitiveDependencies>
+      <useTransitiveFiltering>false</useTransitiveFiltering>
+      <includes>
+		<include>org.xerial.snappy:snappy-java</include>
+        <include>org.apache.hive:hive-jdbc</include>
+        <include>org.apache.hive:hive-exec</include>
+        <include>org.apache.hive:hive-service</include>
+        <include>org.apache.hive:hive-metastore</include>
+        <include>org.apache.hive.shims:hive-shims-common</include>
+        <include>org.apache.hive.shims:hive-shims-0.23</include>
+        <include>org.apache.hive:hive-common</include>
+        <include>org.apache.thrift:libfb303</include>
+        <include>org.apache.thrift:libthrift</include>
+        <include>org.apache.oozie:oozie-core</include>
+        <include>org.apache.oozie:oozie-client</include>
+        <include>com.yammer.metrics:metrics-core</include>
+        <include>org.apache.sqoop:sqoop</include>
+        <include>org.apache.sqoop.connector:sqoop-connector-tdch</include>
+        <include>dk.brics.automaton:automaton</include>
+        <include>org.apache.pig:pig</include>
+        <include>pentaho:pentaho-hdfs-vfs</include>
+        <include>com.twitter:parquet-hadoop-bundle</include>
+        <include>org.apache.avro:avro</include>
+      </includes>
+      <excludes>
+        <exclude>*:tests:*</exclude>
+      </excludes>
+    </dependencySet>
+  </dependencySets>
+</assembly>
