+++ shims/mapr60/impl/src/main/java/org/apache/hadoop/hive/jdbc/HiveDriver.java
