+++ shims/mapr60/pom.xml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>pentaho-hadoop-shims-list</artifactId>
+        <groupId>org.pentaho</groupId>
+        <version>8.1.0.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>pentaho-hadoop-shims-mapr60-reactor</artifactId>
+    <packaging>pom</packaging>
+    <properties>
+        <shim.name>mapr60</shim.name>
+        <!--default folder-->
+        <org.xerial.snappy.version>1.1.1.3</org.xerial.snappy.version>
+        <automaton.version>1.11-8</automaton.version>
+        <org.apache.hive.version>2.1.1-mapr-1710</org.apache.hive.version>
+        <pig.version>0.16.0-mapr-1703</pig.version>
+        <dependency.pig.classifier>h2</dependency.pig.classifier>
+        <org.apache.oozie.version>4.3.0-mapr-1710</org.apache.oozie.version>
+        <org.apache.thrift.version>0.9.3</org.apache.thrift.version>
+        <sqoop.version>1.4.6-mapr-1710</sqoop.version>
+        <org.apache.sqoop.connector.version>1.1-mapr-1707</org.apache.sqoop.connector.version>
+        <org.apache.avro.version>1.8.0</org.apache.avro.version>
+        <!-- pmr folder -->
+        <htrace-core.version>3.1.0-incubating</htrace-core.version>
+        <net.sf.flexjson>2.1</net.sf.flexjson>
+        <org.apache.hbase.version>1.1.8-mapr-1710</org.apache.hbase.version>
+        <zookeeper.version>3.4.5-mapr-1710</zookeeper.version>
+        <!--impl and hbase-comparators folders-->
+        <org.apache.hadoop.version>2.7.0-mapr-1710</org.apache.hadoop.version>
+    </properties>
+    <modules>
+        <module>pmr</module>
+        <module>default</module>
+        <module>impl</module>
+        <module>hbase-comparators</module>
+        <module>assemblies</module>
+    </modules>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.apache.hadoop</groupId>
+                <artifactId>hadoop-common</artifactId>
+                <version>${org.apache.hadoop.version}</version>
+                <scope>provided</scope>
+                <exclusions>
+                    <exclusion>
+                        <groupId>*</groupId>
+                        <artifactId>*</artifactId>
+                    </exclusion>
+                </exclusions>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.hadoop</groupId>
+                <artifactId>hadoop-mapreduce-client-core</artifactId>
+                <version>${org.apache.hadoop.version}</version>
+                <scope>provided</scope>
+                <exclusions>
+                    <exclusion>
+                        <groupId>*</groupId>
+                        <artifactId>*</artifactId>
+                    </exclusion>
+                </exclusions>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+</project>
